requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
dj-database-url==2.1.0
Pillow==10.2.0
flake8==7.0.0
pytest==8.1.1
pytest-django==4.8.0
coverage==7.4.3
